// frontend-v2/src/api/types.ts
export type Id = string;
// ...
export interface RoleCapabilities {
  can_view_channels: boolean;
  can_manage_channels: boolean;
  can_manage_roles: boolean;
  can_create_invites: boolean;
  can_send_messages: boolean;
  can_delete_messages: boolean;
  can_attach_files: boolean;
  can_remove_members: boolean;
  can_mute_members: boolean;
  can_connect_voice: boolean;
  can_speak_voice: boolean;
}

export interface ServerRole {
  id: Id;
  server_id: Id;
  name: string;
  can_create_channels: boolean;
  capabilities: RoleCapabilities;
  is_system: boolean;
  position: number;
  member_ids: Id[];
}

export interface Member {
  account_id: Id;
  handle: string;
  identity_pubkey: string;
  has_avatar: boolean;
  display_name?: string;
}